firmware: eliminate double-precision arithmetic from the hot paths
The ESP32-C6 is a 32-bit RISC-V core with a single-precision-only FPU, so
every f64 op is software-emulated. Move all firmware hot-path math to
integer / fixed-point:

- Pattern clock (render loop, runs every loop() pass): epoch is integer ms,
  bit period is integer µs, frame index = (elapsed_ms * 1000) / period_us —
  all integer, no f64. lm_pattern_timing/lm_player_handle FFI take integer
  ms; main.cpp drops its (double)millis()/since/period math.
- Per-LED brightness scale (per LED per frame): stored as Q8 fixed-point;
  scale is (v * q8 + 128) >> 8, an integer multiply+shift.
- Counting-pattern color (polled per LED every pass): block colors are
  pre-reduced to u8 once at set time, so counting_color is pure integer.
- Frame timestamps: sample raw micros() (integer µs), stored/encoded as u32.

The get_frame_timing message is firmware->phone only, so it now carries the
player's native integer units (tMonoUs, bitPeriodUs, patternClockEpochMs;
uint32 so proto3 JSON keeps them numbers). Trace server + phone updated to µs.

The only f64 left in firmware is at the COLD wire encode/decode boundary
(the shared protocol's timestamp/brightness fields are double for the
f64-native phone/Pi and the golden conformance) plus time-sync t0, which is
epoch-scale/fractional and genuinely needs double. None on a hot path. The
pattern generator, arena, and store were already integer/f32.

Full suite (38 bazel targets) green.

+++ firmware/player/tests/phone_client_frames.rs
@@ -104,7 +104,7 @@ fn run_frames_through(player: &mut Player, frames: &[Vec<u8>]) {
         req.decode_from_bytes(bytes)
             .unwrap_or_else(|e| panic!("frame[{i}]: phone frame must decode: {e:?}"));
         let arm = req.r#msg.clone().expect("golden frames always have an arm");
-        let now = 1000.0 + i as f64;
+        let now = 1000 + i as i64;
         let reply = player.handle(req, now, now).map(|r| {
             // Every reply must fit and encode — bounded output is part of
             // the firmware contract.
